# vectispire-angular/openapi.json
{
  "components" : {
    "schemas" : {
...
      "SshKeyCreateRequest" : {
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "private_key" : {
            "type" : "string"
          },
          "public_key" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SshKeySummary" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "encryptionState" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "publicKey" : {
            "type" : "string"
          },
          "usedByRepositories" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "AgentKeyAuth" : {
        "description" : "Pre-shared authentication key for remote scanning agents",
        "in" : "header",
        "name" : "X-Agent-Key",
        "type" : "apiKey"
      },
      "ApiKeyAuth" : {
        "description" : "Programmatic API Key for CI/CD and automation integrations",
        "in" : "header",
        "name" : "X-API-Key",
        "type" : "apiKey"
      },
      "BearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "JWT Session Bearer token obtained via /api/v1/auth/login",
        "name" : "BearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "name" : "Asmolabs Security Team",
      "url" : "https://github.com/asmolabs/vectispire"
    },
    "description" : "Automated supply-chain security, SBOM analysis, vulnerability triage, attack surface inventory, and compliance governance platform.",
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Vectispire Control Plane REST API",
    "version" : "4.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/ssh-keys" : {
      "get" : {
        "operationId" : "list_3",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SshKeySummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ssh-keys-controller" ]
      },
      "post" : {
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SshKeyCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SshKeySummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ssh-keys-controller" ]
      }
    },
    "/api/v1/ssh-keys/{id}" : {
      "delete" : {
        "operationId" : "remove_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "ssh-keys-controller" ]
      }
    },
...
  },
  "security" : [ {
    "BearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Current Vectispire Instance",
    "url" : "/"
  } ],
...
}